+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-Fix: Perf problem with `Carousel` on mobile devices 
+## [2.123.2] - 2021-03-03
+
+Fix: 
+
+- Perf problem with `Carousel` on mobile devices 
 
 ## [2.123.1] - 2021-03-01